feat(profile): fractional age from date of birth for age-sensitive scores
ProfileStore already stores the DOB as the source of truth; add a
fractional ageYears (days / 365.2425) alongside the whole-year age, and
feed it into every age-sensitive compute (Fitness Age, HR zones, VO2max,
Rhythm Age) via UserProfile.age. Age advances continuously instead of
bumping once a year; the integer age stays the display value.

+++ android/app/src/main/java/com/noop/analytics/IntelligenceEngine.kt
@@ -1048,7 +1048,7 @@ object IntelligenceEngine {
             val activitySamples = rhythmSamples.mapNotNull { s ->
                 s.dynAccelG?.let { uniffi.whoop_ffi.ActivitySample(s.ts, it) }
             }
-            RustScores.rhythmAge(activitySamples, tzOffsetSeconds, profile.age.toDouble(), sexInput)?.let { ra ->
+            RustScores.rhythmAge(activitySamples, tzOffsetSeconds, profile.age, sexInput)?.let { ra ->
                 repo.upsertMetricSeries(listOf(MetricSeriesRow(deviceId = computedId,
                     day = saturdayKeyOnOrBefore(newestDay), key = "rhythm_age", value = ra.cosinorAgeYears)))
             }
